split image formats into it's own modules
for now it still only supports the default esp8266 and esp32 families, but this paves the way for adding support for alternative formats later on
