+++ espflash/src/chip/mod.rs
@@ -1,15 +1,12 @@
-use bytemuck::{bytes_of, Pod, Zeroable};
 use strum_macros::Display;
 
 use crate::{
-    elf::{update_checksum, CodeSegment, FirmwareImage, RomSegment},
-    error::{ChipDetectError, FlashDetectError},
+    elf::{FirmwareImage, RomSegment},
+    error::ChipDetectError,
     flash_target::{Esp32Target, Esp8266Target, FlashTarget, RamTarget},
-    flasher::{FlashSize, SpiAttachParams},
+    flasher::SpiAttachParams,
     Error, PartitionTable,
 };
-
-use std::io::Write;
 
 pub use esp32::Esp32;
 pub use esp32c3::Esp32c3;
@@ -21,14 +18,13 @@ mod esp32c3;
 mod esp32s2;
 mod esp8266;
 
-const ESP_MAGIC: u8 = 0xE9;
-const WP_PIN_DISABLED: u8 = 0xEE;
-
 pub trait ChipType {
     const CHIP_DETECT_MAGIC_VALUE: u32;
     const CHIP_DETECT_MAGIC_VALUE2: u32 = 0x0; // give default value, as most chips don't only have one
 
     const SPI_REGISTERS: SpiRegisters;
+
+    fn addr_is_flash(addr: u32) -> bool;
 
     /// Get the firmware segments for writing an image to flash
     fn get_flash_segments<'a>(
@@ -36,8 +32,6 @@ pub trait ChipType {
         bootloader: Option<Vec<u8>>,
         partition_table: Option<PartitionTable>,
     ) -> Box<dyn Iterator<Item = Result<RomSegment<'a>, Error>> + 'a>;
-
-    fn addr_is_flash(addr: u32) -> bool;
 }
 
 pub struct SpiRegisters {
@@ -78,19 +72,6 @@ impl SpiRegisters {
     pub fn miso_length(&self) -> Option<u32> {
         self.miso_length_offset.map(|offset| self.base + offset)
     }
-}
-
-#[derive(Copy, Clone, Zeroable, Pod)]
-#[repr(C)]
-struct ExtendedHeader {
-    wp_pin: u8,
-    clk_q_drv: u8,
-    d_cs_drv: u8,
-    gd_wp_drv: u8,
-    chip_id: u16,
-    min_rev: u8,
-    padding: [u8; 8],
-    append_digest: u8,
 }
 
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Display)]
@@ -162,90 +143,28 @@ impl Chip {
     }
 }
 
-#[derive(Copy, Clone, Zeroable, Pod, Debug)]
-#[repr(C)]
-struct EspCommonHeader {
-    magic: u8,
-    segment_count: u8,
-    flash_mode: u8,
-    flash_config: u8,
-    entry: u32,
+#[derive(Clone, Copy, Debug)]
+pub struct Esp32Params {
+    pub boot_addr: u32,
+    pub partition_addr: u32,
+    pub nvs_addr: u32,
+    pub nvs_size: u32,
+    pub phy_init_data_addr: u32,
+    pub phy_init_data_size: u32,
+    pub app_addr: u32,
+    pub app_size: u32,
+    pub chip_id: u16,
 }
 
-#[derive(Copy, Clone, Zeroable, Pod, Debug)]
-#[repr(C)]
-struct SegmentHeader {
-    addr: u32,
-    length: u32,
-}
-
-// Note that this function ONLY applies to the ESP32 and variants; the ESP8266
-// has defined its own version rather than using this implementation.
-fn encode_flash_size(size: FlashSize) -> Result<u8, FlashDetectError> {
-    match size {
-        FlashSize::Flash256Kb => Err(FlashDetectError::from(size as u8)),
-        FlashSize::Flash512Kb => Err(FlashDetectError::from(size as u8)),
-        FlashSize::Flash1Mb => Ok(0x00),
-        FlashSize::Flash2Mb => Ok(0x10),
-        FlashSize::Flash4Mb => Ok(0x20),
-        FlashSize::Flash8Mb => Ok(0x30),
-        FlashSize::Flash16Mb => Ok(0x40),
-        FlashSize::FlashRetry => Err(FlashDetectError::from(size as u8)),
+impl Esp32Params {
+    pub fn default_partition_table(&self) -> PartitionTable {
+        PartitionTable::basic(
+            self.nvs_addr,
+            self.nvs_size,
+            self.phy_init_data_addr,
+            self.phy_init_data_size,
+            self.app_addr,
+            self.app_size,
+        )
     }
-}
-
-const IROM_ALIGN: u32 = 65536;
-const SEG_HEADER_LEN: u32 = 8;
-
-/// Actual alignment (in data bytes) required for a segment header: positioned
-/// so that after we write the next 8 byte header, file_offs % IROM_ALIGN ==
-/// segment.addr % IROM_ALIGN
-///
-/// (this is because the segment's vaddr may not be IROM_ALIGNed, more likely is
-/// aligned IROM_ALIGN+0x18 to account for the binary file header
-fn get_segment_padding(offset: usize, segment: &CodeSegment) -> u32 {
-    let align_past = (segment.addr - SEG_HEADER_LEN) % IROM_ALIGN;
-    let pad_len = ((IROM_ALIGN - ((offset as u32) % IROM_ALIGN)) + align_past) % IROM_ALIGN;
-    if pad_len == 0 || pad_len == IROM_ALIGN {
-        0
-    } else if pad_len > SEG_HEADER_LEN {
-        pad_len - SEG_HEADER_LEN
-    } else {
-        pad_len + IROM_ALIGN - SEG_HEADER_LEN
-    }
-}
-
-fn save_flash_segment(
-    data: &mut Vec<u8>,
-    segment: &CodeSegment,
-    checksum: u8,
-) -> Result<u8, Error> {
-    let end_pos = (data.len() + segment.data().len()) as u32 + SEG_HEADER_LEN;
-    let segment_reminder = end_pos % IROM_ALIGN;
-
-    let checksum = save_segment(data, segment, checksum)?;
-
-    if segment_reminder < 0x24 {
-        // Work around a bug in ESP-IDF 2nd stage bootloader, that it didn't map the
-        // last MMU page, if an IROM/DROM segment was < 0x24 bytes over the page
-        // boundary.
-        data.write_all(&[0u8; 0x24][0..(0x24 - segment_reminder as usize)])?;
-    }
-    Ok(checksum)
-}
-
-fn save_segment(data: &mut Vec<u8>, segment: &CodeSegment, checksum: u8) -> Result<u8, Error> {
-    let padding = (4 - segment.size() % 4) % 4;
-
-    let header = SegmentHeader {
-        addr: segment.addr,
-        length: segment.size() + padding,
-    };
-    data.write_all(bytes_of(&header))?;
-    data.write_all(segment.data())?;
-
-    let padding = &[0u8; 4][0..padding as usize];
-    data.write_all(padding)?;
-
-    Ok(update_checksum(segment.data(), checksum))
 }
